Tag requirement integration event creation refctor
